// src/api/model/prepareUpdateParams.ts
/**
 * Generated by orval v7.0.1 🍺
 * Do not edit manually.
 * Paratrooper API
 * OpenAPI spec version: 1.0.0
 */
import type { PrepareUpdateParamsExpoAppConfig } from "./prepareUpdateParamsExpoAppConfig";
import type { StorageObject } from "./storageObject";

export type PrepareUpdateParams = {
  channel?: string;
  expoAppConfig?: PrepareUpdateParamsExpoAppConfig;
  fileMetadata: StorageObject[];
  message: string;
  runtimeVersion: string;
};