// src/api/model/updateMetadata.ts
/**
 * Generated by orval v7.0.1 🍺
 * Do not edit manually.
 * Paratrooper API
 * OpenAPI spec version: 1.0.0
 */
import type { UpdateMetadataFileMetadata } from "./updateMetadataFileMetadata";

export type UpdateMetadata = {
  bundler: string;
  fileMetadata: UpdateMetadataFileMetadata;
  version: number;
};